MAINT: stats: rename stats/_page_trend_test.py to _page_trend_test_.py
This is pyodide process specific AFAICS, and seems harmless for SciPy itself.
