+++ scipy/stats/__init__.py
@@ -499,7 +499,7 @@ from ._resampling import bootstrap, monte_carlo_test, permutation_test
 from ._entropy import *
 from ._hypotests import *
 from ._rvs_sampling import rvs_ratio_uniforms
-from ._page_trend_test import page_trend_test
+from ._page_trend_test_ import page_trend_test
 from ._mannwhitneyu import mannwhitneyu
 from ._fit import fit
 from ._covariance import Covariance
