+++ scipy/stats/_page_trend_test_.py
